refactor: moved styles to Upload/styles/
