+++ iqps-frontend/src/components/Upload/styles/file_card.scss
@@ -1,4 +1,4 @@
-@import "../../styles/variables";
+@import "../../../styles/variables";
 
 .file-card {
 	display: flex;
