+++ iqps-frontend/src/components/Upload/styles/upload_instructions.scss
@@ -1,3 +1,5 @@
+@import "../../../styles/variables";
+
 .upload-instructions {
 	width: 30%;
 	height: 100%;
